Add test for role selection
