+++ src/components/profile/AdminUserManager.test.tsx
@@ -1,9 +1,10 @@
 import * as React from "react";
 import { range } from "lodash";
-import { getUsers } from "../../api/api";
+import { getUserEtag, getUsers, updateUser } from "../../api/api";
 import { Account } from "../../model/account";
 import { renderWithUserContext } from "../../../test/helpers";
 import AdminUserManager from "./AdminUserManager";
+import { fireEvent, waitFor } from "@testing-library/react";
 jest.mock("../../api/api");
 
 const users: Array<Partial<Account>> = range(0, 15).map(id => ({
@@ -35,4 +36,26 @@ it("renders all users", async () => {
             expect(queryByText(user.email!)).toBeInTheDocument();
         }
     });
+});
+
+test("role update selection", async () => {
+    const user = users[0];
+    const newRole = "cidc-admin";
+    getUserEtag.mockResolvedValue("test-etag");
+    updateUser.mockImplementation(async (t, i, e, updates) => {
+        expect(updates.role).toBe(newRole);
+        return { ...user, updates };
+    });
+    const { findAllByText, getByText } = renderWithUserContext(
+        <AdminUserManager />,
+        currentUser
+    );
+    const roleSelect = (await findAllByText(user.role!))[0];
+
+    // Change the user's role to cidc-admin
+    fireEvent.mouseDown(roleSelect);
+    fireEvent.click(getByText(newRole));
+
+    // Update request gets sent
+    waitFor(() => expect(updateUser).toHaveBeenCalled());
 });
