+++ src/components/profile/AdminUserManager.tsx
@@ -1,12 +1,126 @@
 import * as React from "react";
-import { Typography, Card, CardHeader, CardContent } from "@material-ui/core";
-import { getUsers } from "../../api/api";
+import {
+    Typography,
+    Card,
+    CardHeader,
+    CardContent,
+    makeStyles,
+    TableCell,
+    Tooltip,
+    FormControl,
+    MenuItem,
+    Select,
+    Switch,
+    Button
+} from "@material-ui/core";
+import { getUserEtag, getUsers, updateUser } from "../../api/api";
 import { Account } from "../../model/account";
-import AdminUserTableRow from "./AdminUserTableRow";
 import { SupervisorAccount } from "@material-ui/icons";
 import PaginatedTable, { ISortConfig } from "../generic/PaginatedTable";
 import { useUserContext } from "../identity/UserProvider";
 import { withIdToken } from "../identity/AuthProvider";
+import { ORGANIZATION_NAME_MAP, ROLES } from "../../util/constants";
+import UserPermissionsDialogWithInfo from "./AdminUserPermissionsDialog";
+
+const useRowStyles = makeStyles(theme => ({
+    disabled: {
+        color: theme.palette.text.secondary
+    }
+}));
+
+interface IAdminUserTableRowProps {
+    user: Account;
+    reloadUsers: () => void;
+}
+
+const AdminUserTableRow: React.FC<IAdminUserTableRowProps> = withIdToken(
+    ({ token, user: userProp, reloadUsers }) => {
+        const [user, setUser] = React.useState<Account>(userProp);
+        const [openPermsDialog, setOpenPermsDialog] = React.useState<boolean>(
+            false
+        );
+
+        const classes = useRowStyles();
+        const cellClass = user.disabled ? classes.disabled : undefined;
+
+        const doUserUpdate = (updates: Parameters<typeof updateUser>[3]) => {
+            getUserEtag(token, user.id).then(etag => {
+                updateUser(token, user.id, etag, updates).then(updatedUser => {
+                    setUser(updatedUser);
+                    reloadUsers();
+                });
+            });
+        };
+
+        return (
+            <>
+                <TableCell>
+                    <Tooltip
+                        title={
+                            user.disabled
+                                ? "enable this account"
+                                : "disable this account"
+                        }
+                    >
+                        <Switch
+                            size="small"
+                            color="primary"
+                            checked={!user.disabled}
+                            onChange={() =>
+                                doUserUpdate({ disabled: !user.disabled })
+                            }
+                        />
+                    </Tooltip>
+                </TableCell>
+                <TableCell className={cellClass}>{user.email}</TableCell>
+                <TableCell className={cellClass}>
+                    {user.first_n} {user.last_n}
+                </TableCell>
+                <TableCell className={cellClass}>
+                    {ORGANIZATION_NAME_MAP[user.organization]}
+                </TableCell>
+                <TableCell>
+                    <FormControl
+                        style={{ minWidth: 120, marginRight: 20 }}
+                        disabled={user.disabled}
+                    >
+                        <Select
+                            value={user.role || ""}
+                            onChange={e => {
+                                doUserUpdate({
+                                    role: e.target.value as Account["role"]
+                                });
+                            }}
+                        >
+                            {ROLES.map(role => (
+                                <MenuItem value={role} key={role}>
+                                    {role}
+                                </MenuItem>
+                            ))}
+                        </Select>
+                    </FormControl>
+                </TableCell>
+                <TableCell>
+                    <Button
+                        size="small"
+                        variant="outlined"
+                        color="primary"
+                        disabled={!user.approval_date}
+                        onClick={() => setOpenPermsDialog(true)}
+                    >
+                        Edit Data Access
+                    </Button>
+                    <UserPermissionsDialogWithInfo
+                        open={openPermsDialog}
+                        grantee={user}
+                        token={token}
+                        onCancel={() => setOpenPermsDialog(false)}
+                    />
+                </TableCell>
+            </>
+        );
+    }
+);
 
 const ADMIN_TABLE_PAGE_SIZE = 15;
